test: rename 3 manual GTK demo runners out of pytest's discovery path
test_{textbox,selectview,selectdialog}.py are interactive demos, not
tests - pytest silently collected them with zero real assertions.
Renamed to demo_*.py with a docstring noting how to run them manually.
